+++ example/verify_methods/verify_all/verify_all_test.dart
@@ -8,6 +8,7 @@ void main() {
     Approvals.verifyAll(
       items,
       processor: (item) => 'Item: $item',
+      options: Options(deleteReceivedFile: true),
     );
   });
 }