+++ example/verify_methods/verify_all_combinations/verify_all_combinations_test.dart
@@ -11,6 +11,7 @@ void main() {
     Approvals.verifyAllCombinations(
       inputs,
       processor: (combination) => 'Combination: ${combination.join(", ")}',
+      options: Options(deleteReceivedFile: true),
     );
   });
 }