+++ example/verify_methods/verify_as_json/json_test.dart
@@ -11,6 +11,7 @@ void main() {
       };
       Approvals.verifyAsJson(
         complexObject,
+        options: Options(deleteReceivedFile: true),
       );
     });
   });