+++ example/verify_methods/verify_query/verify_db_query_test.dart
@@ -1,31 +1,11 @@
-import 'dart:async';
-
 import 'package:approval_tests/approval_tests.dart';
 
-class DatabaseRequestQuery implements ExecutableQuery {
-  final String userId;
-
-  DatabaseRequestQuery(this.userId);
-
-  @override
-  String getQuery() => 'SELECT * FROM users WHERE id = $userId';
-
-  @override
-  Future<String> executeQuery(String query) async {
-    // Simulate a database response
-    await Future.delayed(Duration(seconds: 1)); // Simulate database latency
-    // Mocked database response for the user details
-    if (userId == "1") {
-      return '{"id": "1", "name": "John Doe", "email": "john@example.com"}';
-    } else {
-      return 'Error: User not found';
-    }
-  }
-}
+import '../../../test/queries/db_request_query.dart';
 
 void main() async {
   var dbQuery = DatabaseRequestQuery("1");
   await Approvals.verifyQuery(
     dbQuery,
+    options: Options(deleteReceivedFile: true),
   );
 }