+++ example/verify_methods/verify_sequence/verify_sequence_test.dart
@@ -7,6 +7,7 @@ void main() {
 
     Approvals.verifySequence(
       sequence,
+      options: Options(deleteReceivedFile: true),
     );
   });
 }