+++ lib/approval_tests.dart
@@ -20,3 +20,4 @@ part 'src/reporter/doesnt_match_exception.dart';
 part 'src/core/utils/extensions.dart';
 part 'src/core/utils/converter.dart';
 part 'src/core/utils/executable_query.dart';
+part 'src/namer/approval_number.dart';