+++ lib/src/comparator/command_line_comparator.dart
@@ -13,8 +13,8 @@ final class CommandLineComparator extends ComparatorImp {
     bool isLogError = true,
   }) async {
     try {
-      final String approvedContent = File(approvedPath).readAsStringSync();
-      final String receivedContent = File(receivedPath).readAsStringSync();
+      final String approvedContent = ApprovalUtils.readFile(path: approvedPath);
+      final String receivedContent = ApprovalUtils.readFile(path: receivedPath);
 
       final StringBuffer buffer = StringBuffer("Differences:\n");
       List<String> approvedLines = approvedContent.split('\n');
@@ -22,20 +22,27 @@ final class CommandLineComparator extends ComparatorImp {
 
       int maxLines = max(approvedLines.length, receivedLines.length);
       for (int i = 0; i < maxLines; i++) {
-        if (i >= approvedLines.length || i >= receivedLines.length || approvedLines[i] != receivedLines[i]) {
-          buffer.writeln('${ApprovalUtils.lines(20)} Difference at line ${i + 1} ${ApprovalUtils.lines(20)}');
-          buffer.writeln('Approved file: ${i < approvedLines.length ? approvedLines[i] : "No content"}');
-          buffer.writeln('Received file: ${i < receivedLines.length ? receivedLines[i] : "No content"}');
+        if (i >= approvedLines.length ||
+            i >= receivedLines.length ||
+            approvedLines[i] != receivedLines[i]) {
+          buffer.writeln(
+              '${ApprovalUtils.lines(20)} Difference at line ${i + 1} ${ApprovalUtils.lines(20)}');
+          buffer.writeln(
+              'Approved file: ${i < approvedLines.length ? approvedLines[i] : "No content"}');
+          buffer.writeln(
+              'Received file: ${i < receivedLines.length ? receivedLines[i] : "No content"}');
         }
       }
 
       if (buffer.isNotEmpty && isLogError) {
         final String message = buffer.toString();
-        logError(message.trim());
+        logError(exception: message.trim());
       }
-    } catch (e) {
+    } catch (e, st) {
       if (isLogError) {
-        logError('Error during comparison via Command Line. Error: $e');
+        logError(
+            exception: 'Error during comparison via Command Line. Error: $e',
+            stackTrace: st);
       }
       rethrow;
     }