+++ test/constants/lines.dart
@@ -1,0 +1,7 @@
+part of '../approval_test.dart';
+
+/// `_Lines` class contains the lines of different lengths.
+final class _Lines {
+  /// Line of 25 characters
+  static const String lines25 = '=========================';
+}